requirements.txt
numpy==1.24.4
pillow
requests
imagehash
opencv-python-headless
boto3
